requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==3.2
django-environ==0.9.0
djangorestframework==3.14.0
isort==5.11.2
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
